entero bug fixes
Addressed an issues in the if/then statement so ent_stns was correctly formatted as a vector. The corrects issues that popped up later in the code on the table output and plotting.
Formatted lines with correct indentions so the code is easier to follow outside the if/then statments