Changed NLCD reclass categories
The reclass of 2011 NLCD StreamCat categories was adjusted so the raster categories shown in the static map and the summary table are the same.  PctWdWet2011Ws was moved from the 'Other' class to the 'Forest' class.  PctWdWet2011Ws is the percent of watershed area classified as woody wetland land cover (NLCD 2011 class 90) . Also in the function and table open water (NLCD 2011 class 11) is included in "Other" but is kept as it's own category in the raster because it looks weird to map it that way.